compositorclient: add SurfaceByName() and NativeSurface() apis
Re-worked previous implementation a bit following Pierres review
feedback to better map to the existing IDisplay & ISurface classes.

SurfaceByName() is used on the IDisplay class to get the ISurface.
NativeSurface() is used on the ISurface to get the native surface.

These api's are used with Wayland Weston compositor client when
using waylandsink as the video sink in webkit. I've moved the implementation
to the Implementation.h header which is shared between Westeros and Weston.

wpebackend-rdk calls these apis to obtain the surface handle created by
Thunder, which is then passed to waylandsink as an external surface so
that the video surface is correctly created as a sub-surface
of the browser.

Also RPI compositorclient is updated to have a stub implementation.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>